Fix being unable to drag items sometimes

+++ Assets/Scripts/Game/Systems/ProcessItemSwapTetherSystem.cs
@@ -1,5 +1,6 @@
 using Game.Components;
 using Unity.Entities;
+using UnityEngine;
 
 namespace Game.Systems
 {
@@ -25,18 +26,25 @@ namespace Game.Systems
                 }
                 else
                 {
-                    EntityManager.CreateSingleton(tether);
+                    singletonEntity = EntityManager.CreateSingleton(tether);
                     hasSingleton = true;
                 }
 
                 EntityManager.DestroyEntity(entity);
             }).WithoutBurst().WithStructuralChanges().Run();
-
+            
+            // Here we get the singleton entity again because it may have changed in the previous loop
+            hasSingleton = SystemAPI.TryGetSingletonEntity<DragTether>(out singletonEntity);
             if (hasSingleton)
             {
-                Entities.ForEach((Entity entity, in InventorySlotUnhoveredEvent hoveredEvent) =>
+                var tether = SystemAPI.GetSingleton<DragTether>();
+                Entities.ForEach((Entity entity, in InventorySlotUnhoveredEvent unhoveredEvent) =>
                 {
-                    EntityManager.DestroyEntity(singletonEntity);
+                    if (unhoveredEvent.itemView == tether.destinationItemView)
+                    {
+                        EntityManager.DestroyEntity(singletonEntity);
+                    }
+
                     EntityManager.DestroyEntity(entity);
                 }).WithoutBurst().WithStructuralChanges().Run();
             }
